requirements.txt
jupyterlab
pandas
numpy
scikit-learn
sentence-transformers
transformers
torch
Flask
google-genai
python-dotenv
requests